requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
Django==4.1.4
django-environ==0.9.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
psycopg2==2.9.5
redis==4.4.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
